Moved components to folder
